Remove h1 and use lists for entire menu
