+++ app/assets/stylesheets/pages/menu.css.scss
@@ -1,0 +1,95 @@
+.pages, .documents {
+    .menu {
+        margin-bottom: 2em;
+        background-color: $samfundet-light-grey;
+        text-transform: uppercase;
+
+        @include span-columns(12);
+        @include media($tablet-desktop) {
+            border-right: 1px solid #000;
+            background-color: transparent;
+            @include span-columns(3 of 12);
+        }
+
+        .menu-content {
+            display: none;
+
+            @include media($tablet-desktop) {
+                display: block;
+            }
+        }   
+    }
+}
+.menu-content {
+    padding: 0 5px 5px 5px;
+    ul {
+        list-style-type: none;
+        margin: 0;
+        
+        &:not(.menu_root) {
+            @include transition(max-height, 0.7s, ease-in-out);
+            overflow: hidden;
+            opacity: 0.5;
+            max-height: 0px;
+        }
+
+        &.active {
+            opacity: 1;
+            max-height: 1000px;
+        }
+    }
+    li {
+        margin: 10px 0px 0px 15px;
+        font-weight: 200;
+        
+        & > a::before {
+            content: '| ';
+            color: $samfundet-black;
+        }
+        
+        &:not(.submenu) {
+            @include media($mobile) {
+                border-bottom: 1px dotted rgba(0, 0, 0, 0.3);
+            }
+        }
+
+        &:last-child {
+            border-bottom: none;
+        }
+
+        &.submenu {
+            cursor: pointer;
+            color: $samfundet-red;
+            font-weight: 400;
+            
+            &::before {
+                content: "+";
+                display: inline-block;
+                color: $samfundet-black;
+                margin-right: 5px;
+                @include transition(transform, 0.3s);
+            }
+            &.active::before {
+                content: "+";
+                display: inline-block;
+                color: $samfundet-black;
+                margin-right: 5px;
+                @include transform(rotate(405deg));
+            }
+        }
+    }
+
+    a {
+        display: block;
+        color: black;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+        overflow: hidden;
+    }
+
+    a.current {
+        font-weight: 500;
+    }
+}
+
+
